+++ .gitignore
@@ -206,4 +206,6 @@ marimo/_static/
 marimo/_lsp/
 __marimo__/
 gemini.sh
+chutes.sh
+lms.sh
 data
